requirements.txt
Flask == 1.1.2
pyTelegramBotAPI == 3.7.6
requests == 2.25.1
mysql-connector-python == 8.0.23